Feature/unify build (#14)
Unify the python and fortran source builds under one source tree and make system.

This PR overhauls the treatment of the fortran sources in the build process. Previously the python wrapper image built on top of an existing compiled Fortran image. While this did separate concerns, it had the drawback that changing the Fortran source would require recompiling all of the Fortran to rebuild the docker image in order to test the impact of the changes on the Python wrapper. This was way too long of a turn-around time.

Instead, the legacy model sources now live as a git submodule under `/lib/FV3`. `/lib` is the standard directory for Python packages used for compiled sources. The makefile in `/lib` is set up to call the makefile for the legacy Fortran model to produce build artifacts. This makefile in turn is called during the build for the python sources. As a result, `python setup.py build_ext --inplace` will update only the necessary Fortran sources using make. This plays well with our docker dev environment setup.

Some API/design changes:
- build_32 has been removed. The short reason for this is the previous mechanism for telling Python to use the "32-bit" (at least for dycore) Fortran model was removed (we no longer build off of a compiled Fortran image), and having a "32-bit" wrapper is not a high priority. We could add this in later if we want.
- The dockerfile is now multi-stage, to mirror the structure used in the Fortran repo.
- The Fortran sources are retrieved using an ssh key.
- Inside the image, the sources now live in `/fv3gfs-python/`. Previously this was `/cython_wrapper/`.
- The `python` shortcut for `python3` has been removed, since we have plans to use google cloud utilities that install python 2 as a dependency.
- FV3GFS_BUILD_DIR no longer needs to be set, since it is a static directory under `lib/`. It is still set in a few places for convenience.
- `python3` is called explicitly in make files.
- Docker images are now prepended with our GCR url.
- The legacy restart test was removed, since this functionality (restarting the model using the legacy restart routines) was broken and not something we want to support.
- Input data will once again be cached in `/inputdata`

Some changes were done to make tests pass, now that Circle CI is set up.
- Fixed variable typo in `fv3gfs/_ghost_cells.py` that caused an exception
- Correctly prepend the directory to the filename for reading legacy restart initial time

Commit history:

* Added readme for external directory

* use key to retrieve fortran submodule

* refer to fv3gfs fortran code in submodule

* update fv3 to latest master

* updated fv3config to latest version

* use make -C for style consistency

* append input directory name to filename for coupler.res

* use python3

* hardcode fv3gfs build path now that we've set the file structure

* fix typo in value

* remove legacy restart test (not supported functionality)

* update fortran code to latest version

* compile sources with -j8

* remove token from submodule url

* added ssh fingerprint for cloning fv3gfs fortran code

* add fingerprint for default build also

* fix whitespace

* pull environment image

* remove 32bit build (which was not 32bit anyways)

* use multi-stage docker build

* update fv3config to latest commit

* remove reference to unused remote environment image

* remove print statement

* updated readme docker instructions

* add nano to environment

* capitalized Makefile and added PHONY targets

* capitalize Makefile in docker commands

* specify fv3config cache dir for image