configure circleci to update a gh-pages branch with docs (#48)
This PR adds automatic build and deployment of the docs to github pages on CircleCI. Any merge or push to master will trigger a rebuild. The domain is publicly available, but is hidden behind a random string. The idea is that it's not really bad if the docs were to get read by someone who wouldn't normally have permissions, but that we don't want to be publicizing them until the package is ready for release.

Commit history:

* configure circleci to update a gh-pages branch with docs

* fix docs_deploy workflow name

* add ssh fingerprint with push permissions

* fix ssh fingerprint for gh-pages pushing

* hide docs behind pseudorandom folder

* fix gh-pages url

* remove extraneous https from hyperlink

* remove viewcode module

* temporarily remove norobots directive from index

* re-enable norobots, disable jekyll in docs

* include dotfiles

* only build docs on master

* fix _static location in conf

* fix hyperlink style to markdown