Re-merge fv3util into fv3gfs-python (#97)
The build system for fv3gfs-python is presently too coupled with fv3util to treat them as independent packages. This coupling must be removed before we could consider separating them into independent repositories. This PR moves fv3util back into fv3gfs-python to ensure they are updated in tandem.

I've done this as a re-merge rather than a revert to preserve the six PRs that occurred on fv3util when it was in a separate repository.

The only changes are to move the files in fv3util from its submodule into the parent repository. If there are subsequent changes needed they should be done in a follow-on PR, so the file differences can be evaluated (right now, every file and line in fv3util is included in the diff).

Subsequent changes needed:
- Update release instructions to use bumpversion independently for fv3util and fv3gfs-python packages