Add jenkins CI tests for fv3util (#101)
We'd like to test that functionality works on Piz Daint. This PR adds fv3util unit tests to Jenkins which run on Piz Daint.

- added a "cpu only" test script in the jenkins folder and configured Circle CI to use this script. Configured Jenkins to run this test on commit to `master` or `develop` from the head node. However, I also disabled this test on Jenkins since the "gpu" test also runs these tests. A later PR will re-enable this test and disable the CPU tests in the "gpu" job. This is the "fv3util-cpu" test on Jenkins.
- added a "gpu" test script in the jenkins folder and configured Jenkins to run this test on commit to `master` or `develop` branches (can be changed in Jenkins control panel), in a GPU job, or when "launch jenkins" is put as a message on a PR. These are the "fv3util-gpu" and "fv3util-gpu-pr" tests on Jenkins.

The test does not pass yet. Fixing the code and some optimizations I want to make to have the tests run faster (in parallel) are left for future PRs, specifically #109