Add fast and gpu-only options for fv3util tests (#113)
This PR adds flags to the fv3util pytest tests.

We'd like to be able to run only test that use the GPU on Piz Daint's gpu nodes. This PR adds a `--gpu-only` flag which can be used with pytest to run only those tests.

For rapid development, I appreciate being able to periodically run quick tests to check if I've broken existing functionality. The current unit tests take 3+ minutes to run in full, due to a large number of parameter settings which must be tested. This PR adds a `--fast` flag to run only a subset of these options, which takes about a minute to run in total.