Simplify python environment specification (#122)
The development environment has a different version of `black` and `flake8` than the CI, which leads to many false positives when running `make lint` locally. This happens because `gt4py` or `dawn4py` override the installed version of black. This was difficult to discover because `pip` is called from many places in the Dockerfile. This situation is even more confusing/difficult to understand because the dockerfile use `apt-get` to install some packages. 

This PR makes the python environment more deterministic using a pattern we've used in `fv3net` couple of times. It ensures that the packages listed in `requirements.txt` are the ones present in the built `fv3gfs-wrapper` image. Almost all python packages are now installed via pip, and `--no-deps` flags are added to all other `pip install` commands to ensure that they do not override the packages listed in `requirements.txt`. 

If one of our packages adds a dependency, we will need to modify `requirements.txt`. But I think the predictability/reproduciblity of this setup is worth it. We should also consider pinning the versions of the gt4py and other libraries.