Getters for gfsphysics diagnostics (#161)
This PR adds some routines for getting physics diagnostics. I used this feature in an emulator sanity check here: https://github.com/VulcanClimateModeling/fv3net/pull/682/. One could also use this to output the fortran diagnostics via python, or to use the physics tendency outputs as an ML feature.

Add two new public functions
- get_diagnostic_by_name
- get_diagnostic_metadata_by_name

Atttempting to get a variable that has not been allocated will cause a segfault (e.g. a 3d physics component if ldiags3d=false). I explored adding a check for this, but unfortunately it is not possible. In Fortran, I can check if a pointer is associated with a target, and if an allocatable array is allocated, but not if a pointer is associated with an allocated array. This same problem exists for the fortran diag_table, so this feature has the same safety guarantees as the normal model does---very few.