automatically transpose quantities in setters (#165)
Previously, it was possible to call a setter with a variable that has the wrong dimension ordering for what is expected in the Fortran code. This PR adds a step in the setters which transposes the Quantity to the expected dimension ordering before setting the Fortran values. This increases the safety of using the setters, and avoids issues where with the new dynamical core we might accidentally set transposed values (since the new dycore uses a different dimension ordering).