Added Jupyter notebook example for random forest (#195)
This PR adds an example using a simple random forest predictor online in the model. Examples are included both for use in a runfile, and investigation in an iPython notebook. The requirements for the random forest predictor (`sklearn-json` and its dependencies) are included as an extras option.

A test is added but not included in CircleCI. Instead the example is pinned to a specific version of the wrapper image, available on our public GCR repo. This is done because the example is meant to be proof-of-concept, meant to be re-implemented by users using our example as a base, instead of used directly as an end product. Running the test takes about 2.5 minutes and would trigger a download of a C48 run directory.

Changes:
- added `fv3gfs.wrapper.examples` with a random forest predictor for use in examples
- removed gt4py from wrapper image, as it is unused and we are not basing off of the wrapper image for fv3core
- added Dockerfile defining image to run a Jupyter notebook in parallel
- added parallel Jupyter notebook example using random forest
- fixed bug where runfile examples would use an older untagged image over the most recently build tagged image
- added typing_extensions to requirements.txt (it should have been added in an earlier PR)