Add test that prescribing SST alters model evolution (#256)
This PR adds a test confirming that with the use of the namelist flag added in VulcanClimateModeling/fv3gfs-fortran#173, setting the "ocean_surface_temperature" modifies the evolution of the model's prognostic variables (e.g. air temperature).

In between VulcanClimateModeling/fv3gfs-fortran#173 and VulcanClimateModeling/fv3gfs-fortran#93, this functionality was broken (see the failing test associated with 5a6febb of this PR). In other words, since VulcanClimateModeling/fv3gfs-fortran#93, Sfcprop%tsfco in the fortran model was modified within the physics driver prior to being used by any physics scheme; therefore when we set it via Python, it was immediately set to something else in the fortran before it could have any impact. The namelist parameter toggles the behavior back to what it was before VulcanClimateModeling/fv3gfs-fortran#93, allowing us to prescribe the SST.