Ensure only one colon in image name for CI push (#297)
Currently the CI rule for pushing the image fails because the
image is named us.gcr.io/vcm-ml/fv3gfs-wrapper:gnu7-mpich314-nocuda:{SHA1}
which is not a valid docker image name (there are two colons). This
PR replaces the second colon with a dash. So the tag will be
gnu7-mpich314-nocuda-{SHA1}

Since the last time the tests for this repo passed, there have been
some changes related to auth for accessing data on the gs://vcm-fv3config
bucket. So there are also some auth-related changes and version
bumps for fv3config/gcsfs/fsspec.

Resolves #296